PararealGPU.jl

Documentation for PararealGPU.jl

PararealGPU.jl

How to use PararealGPU.jl

Constructing the Subproblems

PararealGPU.PararealGPU — Module

Parareal.jl provides functionality to solve an initial value problem in parallel through the use of a predictor-corrector scheme.

PararealGPU.FirstOrderIVP — Type

InitialValueProblem(der, initialValue, domain)

An object representing an initial value problem

PararealGPU.InitialValueProblem — Type

InitialValueProblem

An abstract super-type representing a first- or second-order initial value problem.

PararealGPU.Interval — Type

Interval(lb, ub)

An object with lower and upper bounds.

PararealGPU.Propagator — Type

Propagator(propagator :: Function, discretization :: Int)

Structured representation of a propagation scheme with a discretization.

PararealGPU.SecondOrderIVP — Type

SecondOrderIVP(acceleration, initialPosition, initialVelocity, domain)

An object representing a second-order initial value problem.

PararealGPU.Solution — Type

Solution(domain :: Vector{Float64}, positionSequence :: Vector{Vector{Float64}}, velocitySequence :: Vector{Vector{Float64}})

Structured representation of the solution to a numerical differential equation.

PararealGPU.assignDevices! — Method

assignDevices!(hostVector :: Vector{Host}) :: Nothing

Assign each device to a worker process on the same host.

PararealGPU.correct! — Method

correct!(
subProblemVector        :: Vector{SecondOrderIVP}, 
subSolutionFineVector   :: Vector{Solution}, 
subSolutionCoarseVector :: Vector{Solution},
positionCorrectorVector :: Vector{Vector{Float64}}
)

Correct solutions positions and velocities.

PararealGPU.correctPosition! — Method

correctPosition!(
subSolutionFineVector   :: Vector{Solution}, 
subSolutionCoarseVector :: Vector{Solution},
positionCorrectorVector :: Vector{Vector{Float64}}
)

Correct solutions positions.

PararealGPU.correctVelocity! — Method

correctVelocity!(
subSolutionFineVector   :: Vector{Solution}, 
subSolutionCoarseVector :: Vector{Solution},
velocityCorrectorVector :: Vector{Vector{Float64}}
)

Correct solutions velocities.

PararealGPU.createHostVector — Method

createHostVector(remoteHostNameVector :: Vector{String}, managerVector :: Vector{Int}, devCountVector :: Vector{Int})

Bundle the process IDs with the number of devices on each remote host.

PararealGPU.discretize — Method

discretize(domain :: Interval, discretization :: Int) :: StepRangeLen

TBW

PararealGPU.discretizeKernel! — Method

discretizeKernel!(domainPointVector :: S, step :: T) where {S, T}

Fill discretized domain with middle elements.

PararealGPU.euler — Method

euler(point, slope, step)

The Euler method of numerical integration.

PararealGPU.initializeSubproblems — Method

createSubproblems(ivp :: SecondOrderIVP, discretization :: Int) :: Vector{SecondOrderIVP}

Create subproblems from the given initial value problem and discretization.

PararealGPU.kernel! — Method

kernel!(solver, acceleration, discretizedDomain, position, velocity) :: Nothing

CUDA kernel to propagate the subProblems.

PararealGPU.kernelPrep — Method

kernelPrep(subProblemVector :: Vector{SecondOrderIVP}, discretization :: Int) :: Tuple{Matrix, Array, Array}

Initializes host-based arrays for the discretized domains, positions, and velocities.

PararealGPU.parareal — Method

parareal(ivp :: SecondOrderIVP, coarsePropagator :: Propagator, finePropagator :: Propagator)

Numerically solve the given initial value problem in parallel using a given propagator and discretizations.

PararealGPU.pararealSolution — Method

pararealSolution(solver, acceleration, discretizedDomain, position, velocity) :: Vector{Solution}

Optimize hardware usage and execute the kernel.

PararealGPU.partition — Method

partition(domain :: Interval, discretization :: Int)

Partition an interval into a given number of mostly disjoint sub-domains.

PararealGPU.prepCluster — Method

prepCluster(remoteHostNameVector :: Union{Vector{String}, Int}) :: CachingPool

Prepare a cluster and return a pool of device processes.

PararealGPU.propagate — Function

propagate(ivp :: FirstOrderIVP, propagator :: Propagator, correctors :: Vector{Float64} = zeros(propagator.discretization + 1)) :: Solution

Propagate an initial value problem using a given propagation scheme.

PararealGPU.propagate — Method

propagate(ivp :: SecondOrderIVP, propagator :: Propagator, correctors :: Vector{Float64} = zeros(propagator.discretization + 1)) :: Solution

Propagate an initial value problem using a given propagation scheme.

PararealGPU.propagate — Method

propagate(ivp :: SecondOrderIVP, propagator :: Propagator) :: Solution

Propagate an initial value problem using a given propagation scheme.

PararealGPU.propagateKernel! — Method

propagateKernel!(solver :: F, acceleration :: G, step :: H, positionSequence :: J, velocitySequence :: K) :: Nothing where {F, G, H, J, K}

Propagates on the device.

PararealGPU.showDeviceAssignments — Method

showDeviceAssignments()

Show which device each process has for use.

PararealGPU.spawnManagers — Method

spawnManagers(remoteHostNameVector :: Vector{String}) :: Vector{Int}

Spawn manager processes on each remote host.

PararealGPU.spawnWorkers — Method

spawnWorkers(managerVector :: Vector{Int}) :: Vector{Tuple{String, Int}}

Spawn worker processes that will control device usage.

PararealGPU.symplecticEuler — Method

sympecticEuler(timeStep :: Float64, position :: Vector{Float64}, velocity :: Vector{Float64}, acceleration :: Function) :: Tuple{Vector{Float64}, Vector{Float64}}

Gives the single propagation using the symplectic Euler integrator.

PararealGPU.updateSubproblems! — Method

updateSubproblems!(subProblemVector :: Vector{SecondOrderIVP}, rootSolution :: Solution, acceleration :: Function)

Update the current set of subproblems with the corrected root solution.

PararealGPU.velocityVerlet — Method

velocityVerlet(position :: Vector{Float64}, velocity :: Vector{Float64}, acceleration :: Function, step :: Float64) :: Vector{Float64}

Get the next position and velocity using the Velocity Verlet algorithm.

PararealGPU.PararealGPU

PararealGPU.FirstOrderIVP

PararealGPU.InitialValueProblem

PararealGPU.Interval

PararealGPU.Propagator

PararealGPU.SecondOrderIVP

PararealGPU.Solution

PararealGPU.assignDevices!

PararealGPU.correct!

PararealGPU.correctPosition!

PararealGPU.correctVelocity!

PararealGPU.createHostVector

PararealGPU.discretize

PararealGPU.discretizeKernel!

PararealGPU.euler

PararealGPU.initializeSubproblems

PararealGPU.kernel!

PararealGPU.kernelPrep

PararealGPU.parareal

PararealGPU.pararealSolution

PararealGPU.partition

PararealGPU.prepCluster

PararealGPU.propagate

PararealGPU.propagate

PararealGPU.propagate

PararealGPU.propagateKernel!

PararealGPU.showDeviceAssignments

PararealGPU.spawnManagers

PararealGPU.spawnWorkers

PararealGPU.symplecticEuler

PararealGPU.updateSubproblems!

PararealGPU.velocityVerlet