PararealGPU.jl

Documentation for PararealGPU.jl

PararealGPU.jl

How to use PararealGPU.jl

Constructing the Subproblems

PararealGPU.PararealGPU — Module

Parareal.jl provides functionality to solve an initial value problem in parallel through the use of a predictor-corrector scheme.

PararealGPU.FirstOrderIVP — Type

InitialValueProblem(der :: Function, initialValue :: Number, domain :: Interval)

An object representing an initial value problem

PararealGPU.InitialValueProblem — Type

InitialValueProblem

An abstract super-type representing a first- or second-order initial value problem.

PararealGPU.Interval — Type

Interval{T <: AbstractFloat}(lb :: T, ub :: T)

An object with lower and upper bounds.

PararealGPU.Propagator — Type

Propagator(propagator :: Function, discretization :: Int)

Structured representation of a propagation scheme with a discretization.

PararealGPU.SecondOrderIVP — Type

SecondOrderIVP{T <: AbstractFloat}(acceleration, initialPosition, initialVelocity, domain)

An object representing a second-order initial value problem.

PararealGPU.Solution — Type

Solution(domain :: Vector{T}, positionSequence :: Vector{Vector{T}}, velocitySequence :: Vector{Vector{T}})

Structured representation of the solution to a numerical differential equation.

PararealGPU.assignDevices! — Method

assignDevices!(hostVector :: Vector{Host}) :: Nothing

Assign each device to a worker process on the same host.

PararealGPU.batchProblems — Method

batchProblems(problemVector, my_worker_pool)

TBW

PararealGPU.correct! — Method

correct!(
    subSolutionFineVector   :: Vector{Solution{T}}, 
    subSolutionCoarseVector :: Vector{Solution{T}},
    positionCorrectorVector :: Vector{Vector{T}},
    velocityCorrectorVector :: Vector{Vector{T}}
) :: Nothing where T <: AbstractFloat

Correct solutions positions and velocities.

PararealGPU.createHostVector — Method

createHostVector(remoteHostNameVector :: Vector{String}, managerVector :: Vector{Int}, devCountVector :: Vector{Int})

Bundle the process IDs with the number of devices on each remote host.

PararealGPU.discretize — Method

discretize(domain :: Interval, discretization :: Int) :: StepRangeLen

TBW

PararealGPU.distribute — Method

distribute!(
devPool          :: CachingPool, 
coarsePropagator :: Propagator, 
finePropagator   :: Propagator,
problemVector    :: Vector{SecondOrderIVP}
) :: Vector{Solution}

The parallelization scheme to use from the director process.

PararealGPU.euler — Method

euler(point, slope, step)

The Euler method of numerical integration.

PararealGPU.getRelativeChange — Method

getrelativeChange(old_seq :: Vector{Vector{T}}, new_seq :: Vector{Vector{T}}) :: T where T <: AbstractFloat

Calculate the relative change between two vector sequences.

PararealGPU.getRelativeChange — Method

getRelativeChange(old_vec :: Vector{T}, new_vec :: Vector{T}) :: T where T <: AbstractFloat

Calculate the relative change between two vectors. Results near zero when close, and 1 when far. Bounded between 0 and 1.

PararealGPU.gpu — Method

gpu(
    problemVector  :: Vector{SecondOrderIVP},
    prop           :: Propagator
) :: Vector{Solution}

The parallelization scheme to use from the worker processes.

PararealGPU.hasConverged — Method

hasConverged(
    old_sol :: Solution{T}, 
    new_sol :: Solution{T}, 
    seq :: Symbol; 
    threshold :: T = convert(T, 1.0e-10)
) :: Tuple{Bool, T} where T <: AbstractFloat

Determines convergence of the given property given the previous and and current solutions.

PararealGPU.hasConverged — Method

hasConverged(
    oldSolution :: Solution{T}, 
    newSolution :: Solution{T}; 
    threshold = convert(T, 1.0e-10)
) :: Tuple{Bool, T, T} where T <: AbstractFloat

Determines convergence of both position and velocity given the previous and current solutions.

PararealGPU.initializeSubproblems — Method

createSubproblems(ivp :: SecondOrderIVP, discretization :: Int) :: Vector{SecondOrderIVP}

Create subproblems from the given initial value problem and discretization.

PararealGPU.kernelPrep — Method

kernelPrep(
    subProblemVector :: Vector{SecondOrderIVP{T}}, 
    discretization :: Int
) :: Tuple{Matrix{T}, Array{T, 3}, Array{T, 3}} where T <: AbstractFloat

Initializes host-based arrays for the discretized domains, positions, and velocities.

PararealGPU.parareal — Method

parareal(
    ivp              :: SecondOrderIVP{T}, 
    coarsePropagator :: Propagator, 
    finePropagator   :: Propagator;
    threshold = convert(T, 1.0e-10)
) :: Solution{T} where T <: AbstractFloat

Solve the given IVP using the distributed, GPU-based Parareal algorithm.

PararealGPU.pararealSolution! — Method

pararealSolution!(
    int_scheme        :: F, 
    acceleration      :: G, 
    discretizedDomain :: Matrix{T}, 
    pos_seqs          :: Array{T, 3}, 
    vel_seqs          :: Array{T, 3}
) :: Vector{Solution} where {T <: AbstractFloat, F <: Function, G <: Function}

Optimize hardware usage and execute the kernel.

PararealGPU.partition — Method

partition(domain :: Interval, discretization :: Int)

Partition an interval into a given number of mostly disjoint sub-domains.

PararealGPU.prepCluster — Method

prepCluster(remoteHostNameVector :: Union{Vector{String}, Int}) :: CachingPool

Prepare a cluster and return a pool of device processes.

PararealGPU.propagate — Function

propagate(ivp :: FirstOrderIVP, propagator :: Propagator, correctors :: Vector{Float64} = zeros(propagator.discretization + 1)) :: Solution

Propagate an initial value problem using a given propagation scheme.

PararealGPU.propagate — Method

propagate(ivp :: SecondOrderIVP, propagator :: Propagator) :: Solution

Propagate an initial value problem using a given propagation scheme.

PararealGPU.propagate — Method

propagate(ivp :: SecondOrderIVP, propagator :: Propagator, correctors :: Vector{Float64} = zeros(propagator.discretization + 1)) :: Solution

Propagate an initial value problem using a given propagation scheme.

PararealGPU.propagate_gpu! — Method

propagate_gpu!(
    problemCount :: Int,
    dimension    :: Int,
    t_max        :: Int,
    step         :: Float32,
    pos_seqs_dev,
    vel_seqs_dev
) :: Nothing

TBW

PararealGPU.showDeviceAssignments — Method

showDeviceAssignments()

Show which device each process has for use.

PararealGPU.solve — Method

solve(
    nodeVector           :: Vector{String},
    coarseIntegrator     :: Function,
    coarseDiscretization :: Int,
    fineIntegrator       :: Function,
    fineDiscretization   :: Int,
    acc                  :: Function,
    lowerBound           :: T,
    upperBound           :: T,
    initialPosition      :: Vector{T},
    initialVelocity      :: Vector{T};
    addlocal             :: Bool = false,
    threshold            :: T    = convert(T, 10)
) :: Solution{T} where T <: AbstractFloat

Solves the given problem using the given coarse and fine propagators.

PararealGPU.spawnManagers — Method

spawnManagers(remoteHostNameVector :: Vector{String}) :: Vector{Int}

Spawn manager processes on each remote host.

PararealGPU.spawnWorkers — Method

spawnWorkers(managerVector :: Vector{Int}) :: Vector{Tuple{String, Int}}

Spawn worker processes that will control device usage.

PararealGPU.symplecticEuler — Method

sympecticEuler(timeStep :: Float64, position :: Vector{Float64}, velocity :: Vector{Float64}, acceleration :: Function) :: Tuple{Vector{Float64}, Vector{Float64}}

Gives the single propagation using the symplectic Euler integrator.

PararealGPU.updateSubproblems! — Method

updateSubproblems!(subProblemVector :: Vector{SecondOrderIVP}, rootSolution :: Solution, acceleration :: Function)

Update the current set of subproblems with the corrected root solution.

PararealGPU.velocityVerlet — Method

velocityVerlet(position :: Vector{Float64}, velocity :: Vector{Float64}, acceleration :: Function, step :: Float64) :: Vector{Float64}

Get the next position and velocity using the Velocity Verlet algorithm.

PararealGPU.PararealGPU

PararealGPU.FirstOrderIVP

PararealGPU.InitialValueProblem

PararealGPU.Interval

PararealGPU.Propagator

PararealGPU.SecondOrderIVP

PararealGPU.Solution

PararealGPU.assignDevices!

PararealGPU.batchProblems

PararealGPU.correct!

PararealGPU.createHostVector

PararealGPU.discretize

PararealGPU.distribute

PararealGPU.euler

PararealGPU.getRelativeChange

PararealGPU.getRelativeChange

PararealGPU.gpu

PararealGPU.hasConverged

PararealGPU.hasConverged

PararealGPU.initializeSubproblems

PararealGPU.kernelPrep

PararealGPU.parareal

PararealGPU.pararealSolution!

PararealGPU.partition

PararealGPU.prepCluster

PararealGPU.propagate

PararealGPU.propagate

PararealGPU.propagate

PararealGPU.propagate_gpu!

PararealGPU.showDeviceAssignments

PararealGPU.solve

PararealGPU.spawnManagers

PararealGPU.spawnWorkers

PararealGPU.symplecticEuler

PararealGPU.updateSubproblems!

PararealGPU.velocityVerlet